Safari-609.1.11.1/LayoutTests/accessibility/mac/footer-roledescription-expected.txt

This tests that a footer element and content info has the right role at the right time.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS footer1.role is 'AXRole: AXGroup'
PASS footer1.subrole is 'AXSubrole: AXLandmarkContentInfo'
PASS footer1.roleDescription is 'AXRoleDescription: content information'
PASS contentinfo.role is 'AXRole: AXGroup'
PASS contentinfo.subrole is 'AXSubrole: AXLandmarkContentInfo'
PASS contentinfo.roleDescription is 'AXRoleDescription: content information'
PASS footer2.role is 'AXRole: AXGroup'
PASS footer2.subrole is 'AXSubrole: AXFooter'
PASS footer2.roleDescription is 'AXRoleDescription: footer'
PASS footer3.role is 'AXRole: AXGroup'
PASS footer3.subrole is 'AXSubrole: AXFooter'
PASS footer3.roleDescription is 'AXRoleDescription: footer'
PASS footer4.role is 'AXRole: AXGroup'
PASS footer4.subrole is 'AXSubrole: AXFooter'
PASS footer4.roleDescription is 'AXRoleDescription: footer'
PASS successfullyParsed is true

TEST COMPLETE
